Rename app package to net.hlan.sushi
