+++ .github/workflows/release.yml
@@ -196,7 +196,7 @@ jobs:
         uses: r0adkll/upload-google-play@v1
         with:
           serviceAccountJsonPlainText: ${{ env.GOOGLE_PLAY_SERVICE_ACCOUNT_JSON }}
-          packageName: com.sushi.sshclient
+          packageName: net.hlan.sushi
           releaseFiles: app/build/outputs/bundle/release/app-release.aab
           track: ${{ needs.create-release.outputs.play-track }}
           status: completed
